+++ lib/KnightLab_FeatherUtils/src/KnightLab_FeatherUtils.h
@@ -6,5 +6,7 @@
 #define VBATPIN 9
 
 extern float batteryLevel();
+extern float batteryLevelAveraged();
+
 
 #endif